requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage

flake8
python-dotenv>=0.5.1

setuptools~=65.5.0
pyyaml~=6.0
pandas~=1.5.1
numpy~=1.23.4
tqdm~=4.64.1
scikit-learn~=1.1.3
spacy~=3.4.3
torchtext~=0.6.0
flask
indic-nlp-library